Use streamly classifySessionsOf for sessionizing
Replaced the custom sessionizing code with the streamly combinator
classifySessionsOf. Streamly combinator does not fork threads for timeouts, it
maintains a timer using a heap. The new code saves some memory as well.

Also, the streamly combinator will also implement limiting the number of
sessions or event per session to guarantee an upper limit on memory usage and
dynamic adaptation of the timeout.